clarify 'Host/Name in Template' by adding ascii diagram
This update tries to make it really clear 'host' is a key word that exists
in two namespaces 1) as a apply parameter, and 2) template value.  It would
have been nice if these string had been different, but we have what we do
and the best in this situation is to try to be as clear and explicit as
possible the two 'hosts' are separate but related.

Co-authored-with: Jack Holdsworth <jholdsworth@cloudflare.com>
Signed-off-by: Sami Kerola <kerolasa@iki.fi>